Move vmi reset func test to lifecycle tests
