+++ tests/BUILD.bazel
@@ -70,7 +70,6 @@ go_test(
         "pool_test.go",
         "portforward_test.go",
         "replicaset_test.go",
-        "reset_test.go",
         "security_features_test.go",
         "stability_test.go",
         "swap_test.go",
